requirements.txt
google-generativeai==0.8.0
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
python-dotenv==1.0.0
requests==2.31.0
